requirements.txt
requests
python-dotenv
google-cloud-storage
flask
